+++ src/Exceptions/ManipulationException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PdfStudio\Laravel\Exceptions;
+
+use RuntimeException;
+
+class ManipulationException extends RuntimeException {}